requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# FTP_Brute/socks_brute_ftp.py: 4
PySocks == 1.7.1

# FTP_Brute/socks_brute_ftp.py: 7,8
stem == 1.8.0
